move spell special fonts
